Abort on (helper) threads dispatching non-conforming event sequence
Runtimes that don't OMPT-support target constructs have been reported
to create helper threads that lack thread-begin and
implicit-task-begin but dispatch parallel-begin.

In addition, it has been observed that the initial-task creates a
non-undeferred task that gets executed in a parallel region that the
initial-task isn't part of.  This is non-conforming.

See also https://github.com/llvm/llvm-project/issues/62764