[OpenMP] Add oneAPI flag issues to OPEN_ISSUES
Recently, an issues with the oneAPI runtime has been observed
affecting instrumentation with Score-P.

In particular, the OpenMP flags `-fiopenmp` and `-qopenmp` can
cause issues with OMPT. When trying to execute a very small
example containing `#pragma omp sections`, the OpenMP runtime
incorrectly reports a work callback with a work type of `loop`
instead of `sections` at the end of the sections construct. This
causes Score-P to abort with an inconsistent profile. For C / C++,
users can use the `-fopenmp` flag to work around this issue.
Fortran reports the wrong work type regardless of the flag. This
issue has been observed with oneAPI 2023.0.0 and newer but may
also affect older versions.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>