Filter additional outlined OpenMP functions introduced by oneAPI
Function names containing `.extracted` and `_tree_reduce_` are seen
with ifx and icpx with compiler flags `-fiopenmp` and `-qopenmp`, but
not with `-fopenmp`.

Function names containing `DIR.OMP.` are seen with ifx with all
possible OpenMP flags.

Intel might change it's icx and icpx naming scheme to follow the
Fortran one.

See #256