[LLVM Plugin] Implement configure check and basis
This plugin introduces the configure check being used for the LLVM
Plugin to be used in upcoming commits. The plugin will function similar
to the GCC plugin. The configure check will make sure that the plugin
can be correctly linked to a program and checks flags based on the
compiler being used.

To ensure that the plugin will be used instead of the normal
instrumentation flags, the check was modified accordingly. To implement
the adapter, a new build folder is introduced which contains the base
logic.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>