[OMPT] Use `-mp=ompt` explicitly with OMPT & NVHPC
While we use the flag `-mp=ompt` for compiling Score-P, at least for
CMake programs, one could run into the issue where parts of the OMPT
interface were not completely initialized when not passing `-mp=ompt` to
the compile flags again during instrumentation. This commit adds the
flag when NVHPC is used.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>