Add `-Xarch_host` for LLVM + OpenMP target
LLVM based runtimes have linking issues when trying to compile code with
OpenMP target regions because the instrumenter functions do not exist on
target devices. This issue was solved in LLVM 16.0.1 where the flag
`-Xarch_host` can now be used for OpenMP target as well. This commit
adds the flag if the configure check for compiler instrumentation
succeeds.

This change was tested with OneAPI 2023.2.0, aomp 17.0-0, NVHPC 23.7 and
LLVM 16.0.6/17.0-rc1 where it solved the issue and didn't cause new ones.
Older versions like LLVM 15.0.6 still throw the same error message. For
them, we already have a workaround present in OPEN_ISSUES.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>