[OMPT] Handling of taskwait depend
Since `#pragma omp taskwait depend` uses tasks instead of sync regions
for its handling, the event was previously missed by Score-P. With this
commit, handling for the undeferred version is added. Using the
combination `#pragma omp taskwait nowait depend` might dispatch a
non-undeferred task, which isn't handled yet. However, as no supporting
runtime currently dispatches this kind of task, this shouldn't be a problem.

Once runtimes dispatch deferred tasks for taskwait, a warning will be
shown, that this case is not implemented yet.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>