[Configure] Enable HIP if checks succeed
Previously, configure would not enable the HIP adapter even when all
checks were successful. This happened because the default was set to
"no", which is checked against "yes" at several places.

In addition, removed the redundant check of the user selected and
configure result being both no, since this would hide the reason why the
adapter is not available.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>