[OMPT] Fix stored_task for parallel in tasks
Right now, the stored_task is always updated to the last
encountering_task when a parallel_region is ended. However, this causes
issues when this encountering task is a `#pragma omp task` which is
completed afterwards. When we then try to check if overdue events need
to get triggered, it fails because the pointer is inaccurate.

This change will update the stored_task whenever a task is switched,
fixing the issues showing up for parallel regions inside of tasks.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>