[CUDA] Ignore exit event on context changed memcpy
The CUDA adapter did ignore the enter event for memory copies when the
enter event was being delivered on a context changed context (i.e. a
different thread than the original one). This caused inconsistent
profiles, as we did write an exit event for a missing enter event.
Moved the respective condition, so that exit events are properly
ignored as well.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>