[Compiler] Check instrumentation per language
Allow compiler instrumentation to work correctly when a custom
installation e.g. with GCC & LLVM (gcc, clang++, [some Fortran compiler]) is
used. Previously, only the GCC plug-in would work in this case, with the
LLVM plug-in being disregarded due to macro checks. Also check the
instrumentation on a language basis to make sure that the passed flags are
actually meant for that compiler and work.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>