[OMPT] Ensure OMPT init. on device init callback
Beginning with LLVM 18, the runtime will initialize all visible devices
on program start. Fortunately, this is done at a point where we are able
to initialize Score-P. If we don't initialize it for device init. as
well, program execution of a program containing OpenMP target regions and a
CUDA adapter enabled will probably abort. Therefore, also add the ensure
initialized macro for `ompt_callback_device_initialize`.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>