[LLVM Plugin] Fix compile for no filter + verbose
In !430, the filter file argument was changed to allow passing multiple
filter files to the LLVM plug-in. Unfortunately, this also broke passing
verbose without a filter file, as this caused the filter file command
line option to check how many arguments were passed. As we chose to
expect at least one instead of zero arguments, this caused compilation
to abort.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>